requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
supabase
streamlit
requests
pydantic
huggingface-hub
openai
sentence-transformers
pgvector
